Add new dark mode provider and storybook dark mode support to the storybook docs.
