+++ package.json
@@ -26,6 +26,7 @@
         "@storybook/react": "^7.6.17",
         "@storybook/react-vite": "^7.6.17",
         "@storybook/test": "^7.6.17",
+        "@storybook/theming": "^7.6.17",
         "@types/react": "^18.2.56",
         "@types/react-dom": "^18.2.19",
         "@typescript-eslint/eslint-plugin": "^7.0.2",
@@ -37,6 +38,7 @@
         "eslint-plugin-storybook": "^0.8.0",
         "prettier": "^3.2.5",
         "storybook": "^7.6.17",
+        "storybook-dark-mode": "^3.0.3",
         "typescript": "^5.2.2",
         "vite": "^5.1.4"
     },
